+++ unlock-js/CHANGELOG.md
@@ -1,5 +1,9 @@
 # Changes
 
+## 0.3.14
+
+- Adding retries on error 419 (rate limiting)
+
 ## 0.3.13
 
 - Returning the token symbol when retrieving an ERC20 lock.